requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
Jinja2==3.1.3